fix: correct require paths in ci scripts after reorganization
